// .gadget/client/src/models/ShopifyProduct.ts
export const DefaultShopifyProductSelection = {
     __typename: true,
     id: true,
     body: true,
     category: true,
     compareAtPriceRange: true,
     createdAt: true,
     handle: true,
     hasVariantsThatRequiresComponents: true,
     productCategory: true,
     productType: true,
     publishedAt: true,
     shopId: true,
     shopifyCreatedAt: true,
     shopifyUpdatedAt: true,
     status: true,
     tags: true,
     templateSuffix: true,
     title: true,
     updatedAt: true,
     vendor: true
   } as const;
// ...
export type ShopifyProductManager = {
// ...
  view: {
      (query: string, variables?: Record<string, unknown>): Promise<unknown>
      type: 'computedView';
      operationName: 'view';
      gqlFieldName: 'shopifyProductGellyView';
      namespace: null;
      imports: [];
      variables: {
          query: { type: 'String', required: true },
          args: { type: 'JSONObject' }
        };
      variablesType: Record<string, unknown>;
      resultType: Promise<unknown>;
      plan: never;
    }
};